tested + main in comment

+++ ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: vsozonof <vsozonof@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/07 13:29:48 by vsozonof          #+#    #+#             */
-/*   Updated: 2022/11/07 17:37:19 by vsozonof         ###   ########.fr       */
+/*   Updated: 2022/11/09 10:26:49 by vsozonof         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libft.h"
+#include	"libft.h"
 
 size_t	ft_strlen(char *str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i])
